Activity | 1 | Propose a Business Trip
Depicts a proposal for a business trip to Veracruz. It is a business trip and everything is written in English, as explained in this presentation:
It explains the reasons for choosing Veracruz:
Veracruz is a destination that combines history, culture, and business. It stands out for its role as a historic and strategic port for international trade, as well as being a key point in the Gulf of Mexico with connections to America, Europe, and Asia.
It has an international airport, a modern road network, executive hotels, convention centers, and the World Trade Center (WTC), which facilitates travel, conferences, and high-level meetings.
It generates business events such as trade fairs, logistics conferences, trade shows, workshops, and conferences that allow you to establish strategic contacts and stay up to date on global trends.
It offers great benefits, such as learning and professional development, which allows for the creation of business alliances and international networking, thanks to the motivation and creativity it generates for the team, because cultural experiences (Veracruz Carnival, jarocha music, danzón, typical cuisine) are what reinforce integration and productivity.
In conclusion, the trip to Veracruz is not presented as an expense, but as a strategic investment that combines knowledge, contacts, and motivation, leaving both professional and personal benefits